Fixed Test Dependency Version Resolution Conflict
For the longest time I've been dealing with an issue were various dependencies
had conflicting version dependencies for JUint, AndroidX test runner + rules; As
such tests couldn't not compile when the version was updated but the source of
the restricting dependency couldn't be isolated. Looking at them did not reveal
any of libraries to be needing the lower version.

Using a brand new project I managed to adjust the current configuration to
correct the dependencies and surrounding configuration to finally allow using a
newer version of the test dependencies

In the process several unnecessary dependencies have been removed which may have
been restricting dependency resolution and causing the problem. Additionally,
since I have no plan to make use of views in the future I have removed the
dependency on the material and appcompat library and adjusted as needed.

Signed-off-by: Sulaiman Sulaiman <SKSulai98@gmail.com>